funcinl check for go118/go117
